Futureproof addreminder
moved keywordSearch() to /db/index.js in order to make it available for use in other commands
